move conditional check to backup function

+++ install.sh
@@ -27,11 +27,13 @@ create_symlink() {
     ln -sf "$1" "$2/$3"
 }
 
-# Backup file: $1 = path, $2 = filename
-backup_file() {
-    backup_name="$2_$(date +%d-%m-%y_%H%M%S).bak"
-    echo -e " $INFO -> $NORMAL Backing up $2 as '$backup_name'"
-    mv "$1/$2" "$1/$backup_name"
+# Backup file if it exists: $1 = path, $2 = filename
+backup_if_exists() {
+    if [ ! -h "$1/$2" ] && [ -e "$1/$2" ]; then
+      backup_name="$2_$(date +%d-%m-%y_%H%M%S).bak"
+      echo -e " $INFO -> $NORMAL Backing up $2 as '$backup_name'"
+      mv "$1/$2" "$1/$backup_name"
+    fi
 }
 
 setup_local_folders() {
@@ -55,9 +57,7 @@ process_bash_configs() {
     echo -e "$HEADER Processing 'bash configs' $NORMAL"
     config_files=(aliases bash_profile bash_prompt bashrc exports functions inputrc path)
     for file in "${config_files[@]}"; do
-        if [ ! -h "$HOME/.${file}" ] && [ -e "$HOME/.${file}" ]; then
-            backup_file "$HOME" ".${file}"
-        fi
+        backup_file_if_exists "$HOME" ".${file}"
         create_symlink "$SOURCE_DIR/${file}" "$HOME" ".$file"
         # Source newly created files
         if [ "$file" == inputrc ]; then
@@ -71,9 +71,7 @@ process_bash_configs() {
 
 process_misc_configs() {
   echo -e "$HEADER Processing 'tmux config' $NORMAL"
-  if [ ! -h "$HOME/.tmux.conf" ] && [ -e "$HOME/.tmux.conf" ]; then
-      backup_file "$HOME" .tmux.conf
-  fi
+  backup_file_if_exists "$HOME" .tmux.conf
   create_symlink "$SOURCE_DIR/tmux.conf" "$HOME" ".tmux.conf"
 }
 
@@ -87,9 +85,7 @@ process_nvim_configs() {
     # Array since I may break my vim file up into modules at some point
     config_files=(init.vim)
     for file in "${config_files[@]}"; do
-        if [ ! -h "$nvim_config_path/${file}" ] && [ -e "$nvim_config_path/${file}" ]; then
-            backup_file "$nvim_config_path" "$file"
-        fi
+        backup_file_if_exists "$nvim_config_path" "$file"
         create_symlink "$SOURCE_DIR/config/nvim/${file}" "$nvim_config_path" "$file"
     done
     unset config_files
